+++ lib/sepa_rator/builders/credit_transfer/transaction/amount.rb
@@ -8,7 +8,7 @@ module SEPA
           def call
             builder.Amt do
               builder.InstdAmt(
-                format('%.2f', transaction.amount),
+                XmlBuilder.format_amount(transaction.amount),
                 Ccy: transaction.currency
               )
             end